[PAGOPA-2696] fix: updating gitignore
