+++ src/main/java/it/gov/pagopa/fdrtechsupport/controller/OrgResource.java
@@ -1,8 +1,8 @@
-package it.gov.pagopa.fdrtechsupport.resources;
+package it.gov.pagopa.fdrtechsupport.controllers;
 
 import it.gov.pagopa.fdrtechsupport.models.ProblemJson;
-import it.gov.pagopa.fdrtechsupport.resources.response.FrResponse;
-import it.gov.pagopa.fdrtechsupport.resources.response.FrSingleDateResponse;
+import it.gov.pagopa.fdrtechsupport.controllers.response.FrResponse;
+import it.gov.pagopa.fdrtechsupport.controllers.response.FrSingleDateResponse;
 import it.gov.pagopa.fdrtechsupport.service.WorkerService;
 import jakarta.inject.Inject;
 import jakarta.validation.constraints.NotNull;
