+++ src/main/java/it/gov/pagopa/fdrtechsupport/controller/model/request/SearchRequest.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.fdrtechsupport.resources.model;
+package it.gov.pagopa.fdrtechsupport.controllers.model;
 
 import java.util.Optional;
 import lombok.AllArgsConstructor;
