+++ src/main/java/it/gov/pagopa/fdrtechsupport/controller/model/response/FdrFullInfoResponse.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.fdrtechsupport.resources.response;
+package it.gov.pagopa.fdrtechsupport.controllers.response;
 
 import java.time.LocalDate;
 import lombok.*;
