+++ src/test/java/it/gov/pagopa/fdrtechsupport/controller/GenericTest.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.fdrtechsupport.resources;
+package it.gov.pagopa.fdrtechsupport.controllers;
 
 import static io.restassured.RestAssured.given;
